Responded to some of the issues found in the review. Some of them will need further discussion with the team.
Made the skills autocomplete hide when there is nothing in the input box.

Copy and pasting now will trigger the check we have for duplicate entries.

Words that end with an underscore are now trimmed to have the underscore removed.

The inputs are run through a sanitization, basically if someone tries to enter text that will change the css/html of the webpage in some way then it now doesn't work and will display that text as if it was just text.

Moved the message that lets the user know how many characters they can input into a tooltip.

#feature